Down into another rabbit hole
- Go back to `indexmap` and regular `String`, `ustr` is too fancy
- Split `lib.rs` up (too big to fail! Monopolies are bad!)
- Add `clap` for CLI
- Multiple commands: `repl` (what we've been doing) and `run`
